requirements.txt
appdirs==1.4.3
astroid==2.2.5
attrs==19.1.0
black==19.3b0
Click==7.0
future==0.17.1
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.0.0
pkg-resources==0.0.0
pygame==1.9.6
pyglet==1.3.2
pylint==2.3.1
PySDL2==0.9.6
six==1.12.0
toml==0.10.0
typed-ast==1.3.5
wrapt==1.11.1
